packages/external-agents/seedance/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.27.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
